requirements.txt
pynput
python-cfonts
colorama